throttling_sequencer_fastapi/throttling_sequencer/api/graphql/types/game/game_state.py
import strawberry

from throttling_sequencer.api.graphql.types.game.unit import UnitInputType


@strawberry.input
class GameStateInputType:
    player_units: list[UnitInputType]
    enemy_units: list[UnitInputType]
