throttling_sequencer_fastapi/throttling_sequencer/api/http/v1/dto/unit.py
from pydantic import BaseModel

from throttling_sequencer.domain.grid.coordinate import Coordinate


class CoordinateDto(BaseModel):
    x: float
    y: float

    def to_coordinate(self) -> Coordinate:
        return Coordinate(x=self.x, y=self.y)


class UnitDto(BaseModel):
    coordinate: CoordinateDto
    speed: float
    mass: float
    friction: float
